requirements.txt
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mutagen==1.47.0
pycryptodomex==3.20.0
requests==2.32.3
urllib3==1.26.17
websockets==13.0.1
Werkzeug==3.0.4
yt-dlp==2024.8.6
gunicorn
Pillow==10.4.0
